[HWKMETRICS-112] move classes up one level
There is no need for a cassandra-specific package any more since we now only
support one backend.
