+++ core/metrics-core-impl/src/main/java/org/hawkular/metrics/core/impl/CountersMapper.java
@@ -14,34 +14,28 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hawkular.metrics.core.impl.cassandra;
+package org.hawkular.metrics.core.impl;
 
+import java.util.ArrayList;
 import java.util.List;
 
+import com.datastax.driver.core.ResultSet;
+import com.datastax.driver.core.Row;
 import com.google.common.base.Function;
 
-import org.hawkular.metrics.core.api.MetricData;
-import org.joda.time.DateTime;
-import org.joda.time.Duration;
+import org.hawkular.metrics.core.api.Counter;
 
 /**
  * @author John Sanda
  */
-public class ComputeTTL<T extends MetricData> implements Function<List<T>, List<T>> {
-
-    private int originalTTL;
-
-    public ComputeTTL(int originalTTL) {
-        this.originalTTL = originalTTL;
-    }
+public class CountersMapper implements Function<ResultSet, List<Counter>> {
 
     @Override
-    public List<T> apply(List<T> data) {
-        for (T d : data) {
-            Duration duration = new Duration(DateTime.now().minus(d.getWriteTime()).getMillis());
-            d.setTTL(originalTTL - duration.toStandardSeconds().getSeconds());
+    public List<Counter> apply(ResultSet resultSet) {
+        List<Counter> counters = new ArrayList<>();
+        for (Row row : resultSet) {
+            counters.add(new Counter(row.getString(0), row.getString(1), row.getString(2), row.getLong(3)));
         }
-        return data;
+        return counters;
     }
-
 }
