+++ core/metrics-core-impl/src/main/java/org/hawkular/metrics/core/impl/MetricsIndexMapper.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hawkular.metrics.core.impl.cassandra;
+package org.hawkular.metrics.core.impl;
 
 import java.util.ArrayList;
 import java.util.List;
