+++ core/metrics-core-impl/src/test/java/org/hawkular/metrics/core/impl/MetricsITest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hawkular.metrics.core.impl.cassandra;
+package org.hawkular.metrics.core.impl;
 
 import static org.joda.time.DateTime.now;
